+++ arangod/Indexes/Index.cpp
@@ -345,7 +345,7 @@ Index::IndexType Index::type(std::string_view type) {
     return TRI_IDX_TYPE_GEO2_INDEX;
   }
   if (type == "mdi" || type == "zkd") {
-    return TRI_IDX_TYPE_ZKD_INDEX;
+    return TRI_IDX_TYPE_MDI_INDEX;
   }
   if (type == "mdi-prefixed") {
     return TRI_IDX_TYPE_MDI_PREFIXED_INDEX;
@@ -395,7 +395,7 @@ char const* Index::oldtypeName(Index::IndexType type) {
       return iresearch::StaticStrings::ViewArangoSearchType.data();
     case TRI_IDX_TYPE_NO_ACCESS_INDEX:
       return "noaccess";
-    case TRI_IDX_TYPE_ZKD_INDEX:
+    case TRI_IDX_TYPE_MDI_INDEX:
       return "mdi";
     case TRI_IDX_TYPE_MDI_PREFIXED_INDEX:
       return "mdi-prefixed";
