+++ arangod/RocksDBEngine/RocksDBIndex.cpp
@@ -400,7 +400,7 @@ RocksDBKeyBounds RocksDBIndex::getBounds(Index::IndexType type,
       return RocksDBKeyBounds::GeoIndex(objectId);
     case RocksDBIndex::TRI_IDX_TYPE_IRESEARCH_LINK:
       return RocksDBKeyBounds::DatabaseViews(objectId);
-    case RocksDBIndex::TRI_IDX_TYPE_ZKD_INDEX:
+    case RocksDBIndex::TRI_IDX_TYPE_MDI_INDEX:
       return RocksDBKeyBounds::MdiIndex(objectId);
     case RocksDBIndex::TRI_IDX_TYPE_MDI_PREFIXED_INDEX:
       return RocksDBKeyBounds::MdiVPackIndex(objectId);
